requirements.txt
flask>=3.0.0
psycopg2-binary>=2.9.9
google-api-python-client>=2.120.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
httplib2>=0.22.0
pytest>=7.4.0
pytest-flask>=1.3.0
